+++ src/main/java/cn/sd/jrz/autoresource/items/EnergyGeneratorItem.java
@@ -43,11 +43,11 @@ public class EnergyGeneratorItem extends BlockItem {
                     energy = tag.getLong("energy");
                 }
                 if (tag.contains("tickCount", Tag.TAG_LONG)) {
-                    tickCount = tag.getInt("tickCount");
+                    tickCount = tag.getLong("tickCount");
                 }
             }
         }
-        double percent = (int) (tickCount / 20.00 / second * 10000) / 100.00;
+        double percent = (int) (tickCount / 20.00D / second * 10000) / 100.00D;
         tooltip.add(Component.translatable("item.autoresource.energy_generator.tooltip.energy", energy));
         tooltip.add(Component.translatable("item.autoresource.energy_generator.tooltip.output", output));
         if (output < config.getMax()) {
